utils: add fd_copy_full helper
Copying a whole disk image using glib's I/O helpers is unfortunately
rather inefficient, since they do not appear to make use of the kernel's
zero-copy utilities (and they also do a useless poll() on the output fd
before every write()). This adds a sendfile(2) based helper for the
common task of copying from one file descriptor to another until EOF,
with a standard read/write-based fallback.

According to the man page, sendfile() is usable since Linux 2.6.33, so
I'm not even sure it's worth it to add a knob for disabling use of
sendfile if the header is found.