update_handler.c: use fd-based I/O in copy_raw_image
The fact that copy_raw_image ends up picking out the underlying fd from
the given stream, turns a knob in that stream to avoid the fd getting
closed, and that a number of its callers do other operations directly on
that fd, suggests that "raw" fd-based I/O is the proper interface to
use. That also allows using the much more efficient sendfile() system
call when available (which is in practice always on Linux), instead of
doing at least two (to and from userspace), and perhaps more (between
the G*Stream's internal buffers and the stack buffer in
g_output_stream_real_splice) copies of the data.

Somewhat unrelated, this also seems to clean up some leaks in
open_slot_device: the documentation for g_file_new_for_path says that
the return value should be g_object_unref'ed, while the return value
from g_file_get_path should be g_free'd, neither of which seems to
happen.